services/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3
fastapi-users[sqlalchemy]==13.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.1
pgvector==0.3.5
redis==5.0.8
celery==5.4.0
pytest==8.3.3
httpx==0.27.2
urllib3<2
beautifulsoup4==4.12.3
jsonschema==4.23.0
openai>=1.0.0
pyotp==2.9.0
slowapi==0.1.9
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
python-json-logger==2.0.7
